feat(agent-gateway): add a remote-control runtime toggle API
Extract the Remote Control tunnel client into `packages/remote-control` so
both the CLI and the server can drive it, and put an xstate machine in front
of it that owns the off/starting/on/stopping lifecycle.

The server exposes the machine at `GET` and `POST /api/v1/remote-control`,
refusing to start when the bind is not loopback or authentication is
bypassed, and closes the tunnel as part of its shutdown phases.

The relay origin and relay key stay resolved per start, so the toggle reads
the operator's current environment rather than a value captured at boot.
