+++ .changeset/remote-control-toggle-api.md
@@ -1,0 +1,5 @@
+---
+"@pymodel/pythinker-code": minor
+---
+
+The server can now start and stop Remote Control while it runs, through `GET` and `POST /api/v1/remote-control`.
