+++ packages/agent-gateway/src/protocol/error-codes.ts
@@ -73,6 +73,7 @@ export const ErrorCode = {
   EXPERT_TALK_BUSY: 40936,
   EXPERT_TALK_RUN_NOT_RETRYABLE: 40937,
   EXPERT_TALK_CLIENT_UNSUPPORTED: 40938,
+  REMOTE_CONTROL_ALREADY_RUNNING: 40939,
 
   APPROVAL_EXPIRED: 41001,
   QUESTION_EXPIRED: 41002,
