+++ packages/remote-control/package.json
@@ -1,0 +1,50 @@
+{
+  "name": "@pymodel/remote-control",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Pythinker Remote Control tunnel client: registers this machine with the relay and forwards HTTP/WebSocket traffic to the local server.",
+  "license": "MIT",
+  "author": "PyModel",
+  "homepage": "https://github.com/PyModel/pythinker-code/tree/main/packages/remote-control#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/PyModel/pythinker-code.git",
+    "directory": "packages/remote-control"
+  },
+  "bugs": {
+    "url": "https://github.com/PyModel/pythinker-code/issues"
+  },
+  "keywords": [
+    "pythinker",
+    "remote-control",
+    "relay",
+    "tunnel"
+  ],
+  "files": [
+    "dist"
+  ],
+  "type": "module",
+  "imports": {
+    "#/*": "./src/*.ts"
+  },
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "default": "./src/index.ts"
+    }
+  },
+  "scripts": {
+    "build": "tsdown",
+    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "test": "vitest run",
+    "clean": "rm -rf dist"
+  },
+  "dependencies": {
+    "@pymodel/agent-core-v2": "workspace:^",
+    "@pymodel/pythinker-code-oauth": "workspace:^",
+    "ws": "^8.21.3"
+  },
+  "devDependencies": {
+    "@types/ws": "^8.18.0"
+  }
+}
