+++ packages/remote-control/src/index.ts
@@ -1,0 +1,3 @@
+export * from './remote-control';
+export * from './lock';
+export * from './manager';
