+++ packages/remote-control/tsdown.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from 'tsdown';
+
+export default defineConfig({
+  entry: ['./src/index.ts'],
+  format: ['esm'],
+  dts: false,
+  outDir: 'dist',
+  clean: true,
+  deps: {
+    alwaysBundle: [/^@pymodel\//],
+    neverBundle: [],
+  },
+});
