+++ packages/remote-control/vitest.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    name: 'remote-control',
+    include: ['test/**/*.test.ts'],
+    testTimeout: 30_000,
+  },
+});
